added category to answer creation and migrated javascript into a separate component folder
