+++ app/assets/stylesheets/pages/_answernew.scss
@@ -2,6 +2,10 @@
   padding-top: 10vh;
 }
 
+.answer-new-container {
+  padding: 75px
+}
+
 .answer-options-flex {
   display: flex;
 }
