+++ app/controllers/answers_controller.rb
@@ -16,7 +16,7 @@ class AnswersController < ApplicationController
     # binding.pry
     option_ids.each do |option_id|
       option = Option.find(option_id)
-      Answer.create(user: current_user, option: option, date: DateTime.now)
+      Answer.create(user: current_user, option: option, date: DateTime.now, category_id: option.question.category_id)
     end
     redirect_to root_path
   end
